102-magic.c

+++ 0x06-pointers_arrays_strings/101-print_number.c
@@ -7,7 +7,7 @@
  */
 void print_number(int n)
 {
-	unsigned long int x;
+	unsigned int x;
 
 	if (n < 0)
 	{
